+++ test/issuing_purchase_test.exs
@@ -48,6 +48,28 @@ defmodule StarkInfraTest.IssuingPurchase do
     end)
   end
 
+  @tag :issuing_purchase
+  test "update issuing purchase test" do
+    purchases = StarkInfra.IssuingPurchase.query!(limit: 1) |> Enum.take(1)
+    assert purchases != [], "no IssuingPurchase in sandbox"
+    purchase = purchases |> hd
+
+    {:ok, updated_purchase} = StarkInfra.IssuingPurchase.update(purchase.id, description: "Dinner")
+
+    assert updated_purchase.description == "Dinner"
+  end
+
+  @tag :issuing_purchase
+  test "update! issuing purchase test" do
+    purchases = StarkInfra.IssuingPurchase.query!(limit: 1) |> Enum.take(1)
+    assert purchases != [], "no IssuingPurchase in sandbox"
+    purchase = purchases |> hd
+
+    updated_purchase = StarkInfra.IssuingPurchase.update!(purchase.id, description: "Dinner")
+
+    assert updated_purchase.description == "Dinner"
+  end
+
   @content "{\"acquirerId\": \"236090\", \"amount\": 100, \"cardId\": \"5671893688385536\", \"cardTags\": [], \"endToEndId\": \"2fa7ef9f-b889-4bae-ac02-16749c04a3b6\", \"holderId\": \"5917814565109760\", \"holderTags\": [], \"isPartialAllowed\": false, \"issuerAmount\": 100, \"issuerCurrencyCode\": \"BRL\", \"merchantAmount\": 100, \"merchantCategoryCode\": \"bookStores\", \"merchantCountryCode\": \"BRA\", \"merchantCurrencyCode\": \"BRL\", \"merchantFee\": 0, \"merchantId\": \"204933612653639\", \"merchantName\": \"COMPANY 123\", \"methodCode\": \"token\", \"purpose\": \"purchase\", \"score\": null, \"tax\": 0, \"walletId\": \"\"}"
   @signature "MEUCIBxymWEpit50lDqFKFHYOgyyqvE5kiHERi0ZM6cJpcvmAiEA2wwIkxcsuexh9BjcyAbZxprpRUyjcZJ2vBAjdd7o28Q="
   @bad_signature "MEUCIQDOpo1j+V40DNZK2URL2786UQK/8mDXon9ayEd8U0/l7AIgYXtIZJBTs8zCRR3vmted6Ehz/qfw1GRut/eYyvf1yOk="